Fixed a major error
I fixed a major error today, the error would cause the program to
continously loop until you closed the program. I temporarily solved it by
adding removing a feature and setting the total days to days(thus
simulating only one day) it will also print out the day number it's on(up
to the number prior to the amount you entered. Still need to work on some
balancing. I also plan on eventually adding in individualized deer this
wat the simulation will be semi-random in it's outcome.